+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Changes to this project are documented in this file. More detail and links can b
 - AUTOTOOL-1988: Added new Namespace SystemPoller endpoint (/namespace/$namespace/systempoller/$systemOrPoller/$poller?)
 - AUTOTOOL-2056 [GitHub #93](https://github.com/F5Networks/f5-telemetry-streaming/issues/93): Add additional traffic stats for virtual servers and pools
 - AUTOTOOL-1984 [GitHub #85](https://github.com/F5Networks/f5-telemetry-streaming/issues/85): Add support for Splunk proxy options
-- New consumer: DEOS Consumer
+- AUTOTOOL-2134: New consumer: F5_Cloud
 - AUTOTOOL-2028 [GitHub #90](https://github.com/F5Networks/f5-telemetry-streaming/issues/90): Add support for Kafka TLS client authentication
 ### Fixed
 - AUTOTOOL-2089: Fix error where unavailable Custom Endpoint would return HTTP 500